Actual initial commit of code
This first commit is very large because for the most part, this code has
existed in various projects I have worked on in one place or another. So
the real work for this commit was just pulling it together. Also, unit
tests, which didn't exist before because most of the code resided in
setup.py files (which are not usually tested!)