Publish to PyPI; docs; current-version
config.yml: added jobs to publish to PyPI (and test PyPI) on the appropriate
branches. Need to update to do this only when certain files are changed, though.

README.md: Completely overhauled for the new functionality.

commands.py: added current-version field, validation for it, docs on when it
gets used, and now call `utils.get_version_from_git()` with it.

utils.py: rearranged PATCH to be the default; added loging to
`get_version_from_git() to use the fallback current version when there is an
error trying to do git describe. This should be good for creating source distro
packages which will then be turned into wheels by pip on the target machine.

Updated tests for those changes.

Added a CHANGELOG and py.typed, since we are trying to adhere to typing hints.