Unit tests that don't need git present
CHANGELOG.md: looks like we're publishing on the 7th instead

configuration.py: after wrapping the configure() function signature I had to
relocate the pylint disable for the unused arguments.

utils.py: Not re-raising SystemExit, so no need to capture it as a variable.
Replaced SystemExit in git_describe() since we don't actually exit if an error
is raised from this function. Created a base exception and a new exception to
take its place.

test_utils.py: We now mock Popen itself (as a context manager) so we don't
attempt to call the git fetch or git describe during tests. With the old way, we
were, we just ignored its output in favor of our mocked communicate. Now, no
subprocess is actually created, which is more correct and definitely cleaner.